"revert(terraform): validate configuration if triggered by Dependabot (#623)"
The Terraform workflow currently throws an error if triggered by
Dependabot and Dependabot secrets `AZURE_CLIENT_ID`,
`AZURE_SUBSCRIPTION_ID` and/or `AZURE_TENANT_ID` are not configured, as
these secrets are required by the Terraform workflow.

There are two possible solutions to this issue:

1. Set these secrets as not required in the Terraform workflow.
   The downside of this solution is that it can cause confusion regarding
   the usage of the Terraform workflow, as these secrets are still
   required for normal use, they're just not marked as required to
   prevent this error when triggered by Dependabot.

2. Add these Dependabot secrets to the relevant repository and set the
   values to empty strings. This adds an, in my opinion, unnecessary
   and confusing prerequisite for the Terraform workflow.

3. Revert the commit that broke the Terraform workflow.

Going for solution 3 keeps the Terraform workflow as simple as possible,
so that is the solution that I've gone for.

This reverts commit 9d552d9811de33b0ce670c4bddbd39fc083ddc0e.