feat(super-linter): create job summary instead of status checks (#695)
* feat(super-linter): create job summary instead of status checks

Disable creation of status checks per linter. Instead, create a single status check for the entire linter workflow job, then create a job summary that lists all linters that were run.

This is to prevent an overly long list of status checks for pull requests that include many different languages to be linted.

* Update super-linter.yml