feat(terraform): create tag after running `terraform apply`
After running `terraform apply`, create a tag with the given name. This should make it safer to use this workflow to run `terraform apply` from the default branch (instead of from pull request, which we usually recommend) as you can now use tags to keep track of which commits have been deployed.

For example:

- Tag `core@dev` could be used to track which commit has been applied for Terraform configuration `terraform/dev/core`.
- Tag `core@prod` could be used to track which commit has been applied for Terraform configuration `terraform/prod/core`.

Fixes #722