feat: add workflow `terraform-test.yml`
Add a reusable workflow for running infrastructure tests using the Terraform Test framework. Test results are presented in a job summary similar to the one created by the `terraform.yml` workflow.

An almost direct copy of the Terraform Test workflow in [Terraform Baseline](https://github.com/equinor/terraform-baseline/blob/cd1dbe917d4a71a1d3098da13294d5abaa7ec1af/.github/workflows/terraform-test.yml).

The idea is to move that workflow here so that all reusable workflows are collected in a single repository.

When this workflow was originally implemented in the Terraform Baseline repository, the `terraform test` command only supported the `-json` option for displaying results in a format that could be parsed in automation. As a result, this workflow currently uses `jq` to parse the test results and format it in a Markdown table to be displayed in the job summary.

Since then, the `terraform test` command has added support for a new option `-junit-xml` (a standardized format for storing test results), which could potentially replace the custom JSON parsing that we currently do.