feat: disable Git credential persistance (#751)
* feat: disable Git credential persistance

If the value of input `persist-credentials` is set to `true`, the `actions/checkout` action will authenticate Git. The auth token is persisted in the local Git config.

If a workflow does not explicitly need Git to be authenticated, persisting the token poses a potential security risk. By setting the value of this input to `false`, we opt-out of this behavior.

* docs: update best practices

* docs: best practices document should reflect best practices

* docs: fix linter errors