feat(mkdocs): install Python version specified in `.python-version` file (#799)
Support the standard of installing the Python version specified in a
`.python-version` file. Remove action `actions/setup-python` input
`python-version` - the action will then fall back to looking for a
`.python-version` file, else it will use the latest version of Python
available on the runner. You can still specify a version using the
`python_version` input if you want to.