feat(mkdocs): add input `upload_artifact` (#800)
Add an input to control whether or not the MkDocs build artifact should
be uploaded.

For example, use this input to only upload artifact on push to main.
This would allow the workflow to run on PR, to verify that the MkDocs
builds, without taking up GitHub storage space.

Also add corresponding output `artifact_uploaded` that simply outputs
the value of the new input, allowing dependent jobs to run only if an
artifact was uploaded.