ci: enable MkDocs strict mode (#846)
This will cause MkDocs to abort the build on any warnings. Will ensure that all warnings are resolved before merging PRs that update docs.

Requires that all relative paths in the README are replaced with links to the corresponding path in GitHub to mitigate the following warnings, allowing MkDocs to successfully build the site:

```plaintext
Doc file 'index.md' contains a link 'CONTRIBUTING.md', but the target is not found among documentation files.
```