fix(commitlint): workflow fails if triggered by non-push event
Only the push event contains a `head_commit` property in its payload (https://docs.github.com/en/webhooks/webhook-events-and-payloads). As a result, the default value `${{ github.event.head_commit.message }}` for input `message` currently throws an error if this workflow is triggered by a non-push event.

To fix this issue while ensuring backwards compatability with existing versions of this workflow, remove the default value and add the `--last` parameter to the `commitlint` command. The `actions/checkout` will checkout the repository (defaulting to the `main` branch for events that don't specify a ref), then commitlint will lint the last commit message in the checked out repository.

The `fetch-depth` input can be removed from the `actions/checkout` action, since we only need to checkout the last commit (i.e., the commit to be linted).