refactor(terraform): start SSH agent in separate step
Take inspiration from @sefornes and "reduce the magic" by moving the script for starting the SSH agent to a separate step.

The `SSH_PRIVATE_KEY` secret must be set as a job-level environment variable before it can be used in an `if:` conditional (https://docs.github.com/en/actions/how-tos/write-workflows/choose-what-workflows-do/use-secrets#using-secrets-in-a-workflow).