refactor(terraform): generate unique artifact name by default
There could be multiple jobs in a workflow deploying to the same environment, so using the environment name as a unique identifier could fail in some scenarios.

Generate a unique artifact name based on job and run ID instead - there can only be one job of a given ID within a workflow run of a given ID, i.e. the combination of the two will be unique.