feat(python): add input `upload_artifact`
Similar to #800.

Add an input to control whether or not the build artifact should be uploaded.

For example, use this input to only upload the artifact on push to main. This would allow the workflow to run on PR, to verify that checks pass, without taking up GitHub storage space.