feat(databricks-repos): create repo if it doesn't exist
Creating a repo in Databricks requires a Git credential for
authenticating to the Git provider (in our case GitHub). Until now,
we've expected callers of this reusable workflow to create the initial
repo as a prerequisite. However, since we're already creating a Git
credential in this workflow, we might as well check if the repo already
exists, and if it doesn't, try to create it.