feat(release-please-manifest): add output `path_tag_names` (#910)
Add an output `path_tag_names` that contains a JSON object of key-value
pairs, where the key is a path that had a release created, and the value
is the tag name that was created for that release.

The Release Please action dynamically defines outputs for each path that
had a release created. Dynamically defining outputs for a job is not
supported, so we have to convert the entire Release Please action
outputs object to a JSON string, then use jq to construct the static
JSON objects containing the specific outputs that we need.